+++ Changelog.md
@@ -43,6 +43,7 @@
 - Remove unused Javascript files (#7442)
 - Move Docker dependency updates into separate service (#7451)
 - Fixed flaky test due to daylight savings time issue (#7452)
+- Updated Python autotest seed file to illustrate pytest metadata custom markers (#7457)
 
 ## [v2.6.1]